refactor : 구매 금액 타입 long으로 수정

+++ src/main/java/lotto/domain/BuyAmount.java
@@ -3,14 +3,14 @@ package lotto.domain;
 public class BuyAmount {
 
     private static final String DIVISION_ERROR_MESSAGE = "1000원으로 나눌 수 있는 금액만 입력 가능합니다.";
-    private final int buyAmount;
+    private final long buyAmount;
 
-    public BuyAmount(int buyAmount) {
+    public BuyAmount(long buyAmount) {
         validateAbleToDivide(buyAmount);
         this.buyAmount = buyAmount;
     }
 
-    public void validateAbleToDivide(int buyAmount) {
+    public void validateAbleToDivide(long buyAmount) {
         if (buyAmount % 1000 != 0) {
             throw new IllegalArgumentException(DIVISION_ERROR_MESSAGE);
         }
